Output alerts as a stream of events
This commit changes JSON output to contain a single JSON object
per event. Before we had a single JSON object per single API
reponse, which then contained a list of events. None of the
information outside of the event list where relevant to the user,
and the list was harder to parse and harder to use in systems
like splunk etc.

Fixes #65